* applying patch 11 contributed from S. Held * add protocol configuration to OCG XML file
git-svn-id: http://svn.eurecom.fr/openair4G/trunk@6629 818b1a75-f10b-46b9-bf7c-635c3b92a50f

+++ openair2/ENB_APP/enb_config.c
@@ -460,7 +460,7 @@ const Enb_properties_array_t *enb_config_init(char* lib_config_file_name_pP) {
     libconfig_int          tdd_config_s;
     const char*            prefix_type              = NULL;
     libconfig_int          eutra_band;
-    int64_t                downlink_frequency;
+    long long int          downlink_frequency;
     libconfig_int          uplink_frequency_offset;
     libconfig_int          Nid_cell;
     libconfig_int          Nid_cell_mbsfn;
